HTTP視聴に対して認証を追加した
