+++ PeerCastStation/PeerCastStation.HTTP/PeerCastStation.HTTP.csproj
@@ -53,6 +53,7 @@
     <Compile Include="HTTPDummyOutputStream.cs" />
     <Compile Include="HTTPOutputStream.cs" />
     <Compile Include="HTTPSourceStream.cs" />
+    <Compile Include="HTTPUtils.cs" />
     <Compile Include="PlayList.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
